+++ target-root/usr/local/lib/solidgroundux/templates/mod-template.sh
@@ -2,8 +2,8 @@
 # SolidGroundUX - Console Module Template
 # ----------------------------------------------------------------------------------
 # Metadata:
-#   Version     : 1.5
-#   Build       : 2615900
+#   Version     : 1.7
+#   Build       : 2619600
 #   Checksum    : fc132915f0186cf54bc2ac5d3230bd8af461f5db709b6487ee3dbd5a1fca5585
 #   Source      : mod-template.sh
 #   Type        : module